requirements.txt
django>=4.2.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
python-dotenv>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7 